requirements.txt
Django==1.7.6
MySQL-python==1.2.5
celery==3.1.17
django-celery==3.1.16
